20 - Добавить поддержку HttpRange в AttachmentEndpoint

See merge request backstage/components!8

+++ attachments/src/main/java/com/backstage/app/attachment/model/dto/AttachmentDto.java
@@ -40,7 +40,7 @@ public class AttachmentDto
 
 	private String fileName;
 
-	private Integer size;
+	private Long size;
 
 	private String url;
